requirements.txt
azure-storage-blob
azure-ai-formrecognizer
azure-core
azure-search-documents
openai
python-dotenv
fpdf